all hints in
