+++ Slider/Assets/Scripts/Misc/Hints.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2994afe1fefd21b4192b24bd9fb8bfeb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
