--- Slider/Assets/Scripts/Utility/PlayerActionHints.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5dd4c45cbade45940b08d7cc6ad78cd5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
